+++ src/torchmetrics/image/__init__.py
@@ -13,6 +13,7 @@
 # limitations under the License.
 from torchmetrics.image.csi import CriticalSuccessIndex
 from torchmetrics.image.d_lambda import SpectralDistortionIndex
+from torchmetrics.image.d_s import SpatialDistortionIndex
 from torchmetrics.image.ergas import ErrorRelativeGlobalDimensionlessSynthesis
 from torchmetrics.image.mifid import MemorizationInformedFrechetInceptionDistance
 from torchmetrics.image.psnr import PeakSignalNoiseRatio
@@ -31,6 +32,7 @@ from torchmetrics.utilities.imports import (
 
 __all__ = [
     "SpectralDistortionIndex",
+    "SpatialDistortionIndex",
     "ErrorRelativeGlobalDimensionlessSynthesis",
     "PeakSignalNoiseRatio",
     "PeakSignalNoiseRatioWithBlockedEffect",